+++ tests/generic_inline_admin/tests.py
@@ -4,8 +4,8 @@ import warnings
 
 from django.contrib import admin
 from django.contrib.admin.sites import AdminSite
-from django.contrib.contenttypes.generic import (
-    generic_inlineformset_factory, GenericTabularInline)
+from django.contrib.contenttypes.admin import GenericTabularInline
+from django.contrib.contenttypes.forms import generic_inlineformset_factory
 from django.forms.formsets import DEFAULT_MAX_NUM
 from django.forms.models import ModelForm
 from django.test import TestCase, override_settings
