+++ tests/managers_regress/models.py
@@ -2,9 +2,11 @@
 Various edge-cases for model managers.
 """
 
-from django.db import models
-from django.contrib.contenttypes import generic
+from django.contrib.contenttypes.fields import (
+    GenericForeignKey, GenericRelation
+)
 from django.contrib.contenttypes.models import ContentType
+from django.db import models
 from django.utils.encoding import python_2_unicode_compatible, force_text
 
 
@@ -126,7 +128,7 @@ class Child7(Parent):
 # RelatedManagers
 @python_2_unicode_compatible
 class RelatedModel(models.Model):
-    test_gfk = generic.GenericRelation('RelationModel', content_type_field='gfk_ctype', object_id_field='gfk_id')
+    test_gfk = GenericRelation('RelationModel', content_type_field='gfk_ctype', object_id_field='gfk_id')
 
     def __str__(self):
         return force_text(self.pk)
@@ -140,7 +142,7 @@ class RelationModel(models.Model):
 
     gfk_ctype = models.ForeignKey(ContentType)
     gfk_id = models.IntegerField()
-    gfk = generic.GenericForeignKey(ct_field='gfk_ctype', fk_field='gfk_id')
+    gfk = GenericForeignKey(ct_field='gfk_ctype', fk_field='gfk_id')
 
     def __str__(self):
         return force_text(self.pk)
