+++ tests/serializers_regress/models.py
@@ -7,7 +7,9 @@ This class sets up a model for each model field type
 import warnings
 
 from django.db import models
-from django.contrib.contenttypes import generic
+from django.contrib.contenttypes.fields import (
+    GenericForeignKey, GenericRelation
+)
 from django.contrib.contenttypes.models import ContentType
 
 # The following classes are for testing basic data
@@ -109,7 +111,7 @@ class Tag(models.Model):
     content_type = models.ForeignKey(ContentType)
     object_id = models.PositiveIntegerField()
 
-    content_object = generic.GenericForeignKey()
+    content_object = GenericForeignKey()
 
     class Meta:
         ordering = ["data"]
@@ -118,7 +120,7 @@ class Tag(models.Model):
 class GenericData(models.Model):
     data = models.CharField(max_length=30)
 
-    tags = generic.GenericRelation(Tag)
+    tags = GenericRelation(Tag)
 
 # The following test classes are all for validation
 # of related objects; in particular, forward, backward,
